requirements.txt
# fastapi==0.115.0
# uvicorn==0.30.0
# python-multipart==0.0.9
# openai==1.40.0
# anthropic==0.34.0
# elevenlabs==1.4.0
# azure-cognitiveservices-speech==1.40.0
# redis==5.0.8
# python-dotenv==1.0.1
# httpx==0.27.0
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
openai==1.40.0
elevenlabs==1.4.0
python-dotenv==1.0.1